Introduced the Q library
POC for the direction I wish to go down in writing the code. Not happy
with the normal nodejs style as it will lead to hard to read code down
the road. The q style async handling is much more eligant and also will
make it easier to have async completion dependencies where needed in a
clean fashion.