Redesigned the Game Planner Interface
Now game analytics are using a play type (run, pass, screen, etc) as
upper category with playbook as the subcategory (represented as a card
now).

Also added a visual of the play for easier understanding of what
the play in question is and display the breakdown of the good/bad
rankings of opponents using the play against that team.

In addition going back to a modal system so that more information may
be gleemed without having to dig around as much.

Also included the passes into the result set and no longer maintain a
temporary collection set on the backend system to manage the compiled
analytical data.